refactor: Extract LauncherService from 682-line main.swift
Move all process management logic (VHID detection, retry management,
config setup, output bridge, passthru runtime, capture loop) from
main.swift into LauncherService.swift. main.swift is now a 4-line
entry point that creates a LauncherService and calls run().
